+++ plugins/voxelman/block/blockman.d
@@ -3,10 +3,10 @@ Copyright: Copyright (c) 2014-2016 Andrey Penechko.
 License: $(WEB boost.org/LICENSE_1_0.txt, Boost License 1.0).
 Authors: Andrey Penechko.
 */
-module voxelman.core.blockman;
+module voxelman.block.blockman;
 
-import voxelman.core.block;
-import voxelman.core.basicblocks;
+import voxelman.block.block;
+import voxelman.block.basicblocks;
 import voxelman.storage.chunk;
 
 struct BlockMan