+++ plugins/voxelman/block/plugininfo.d
@@ -1,0 +1,6 @@
+module voxelman.block.plugininfo;
+enum id = "voxelman.block";
+enum semver = "0.6.0";
+enum deps = [];
+enum clientdeps = [];
+enum serverdeps = [];